separated section utils from quarter utils
