+++ frontend/src/main/utils/CoursesWithSectionsUtilities.js
@@ -1,0 +1,60 @@
+const location = (loc) => {
+    let ans = "";
+    for (let i = 0; i < loc.length; i++) {
+        ans += `${loc[i].timeLocations[0].building} ${loc[i].timeLocations[0].room}`;
+        if (i + 1 < loc.length) {
+            ans += `,`;
+        }
+    }
+    ans = ans.split(`,`);
+    return ans.map((a)=><div>{a}</div>);
+}
+
+const time = (loc) => {
+    let ans = "";
+    for (let i = 0; i < loc.length; i++) {
+        ans += `${loc[i].timeLocations[0].beginTime}--${loc[i].timeLocations[0].endTime}  ${loc[i].timeLocations[0].days}`;
+        if (i + 1 < loc.length) {
+            ans += `,`
+        }
+    }
+    ans = ans.split(`,`);
+    return ans.map((a)=><div>{a}</div>);
+}
+
+const enroll = (loc) => {
+    let ans = "";
+    for (let i = 0; i < loc.length; i++) {
+        ans += `${loc[i].enrolledTotal}/${loc[i].maxEnroll}`;
+        if (i + 1 < loc.length) {
+            ans += `,`
+        }
+    }
+    ans = ans.split(`,`);
+    return ans.map((a)=><div>{a}</div>);
+}
+
+const instructor = (loc) => {
+    let ans = "";
+    for (let i = 0; i < loc.length; i++) {
+        for (let j = 0; j < Math.min(3, loc[i].instructors.length); j++) { // display 3 instructors at most, or else won't fit in a line
+            ans += `${loc[i].instructors[j].instructor}`;
+            if (j + 1 < Math.min(3, loc[i].instructors.length)) {
+                ans += `, `
+            } 
+        }
+        if (i + 1 < loc.length) {
+            ans += `!`
+        } 
+
+    }
+    ans = ans.split(`!`);
+    return ans.map((a)=><div>{a}</div>);
+}
+
+export {
+    location,
+    time,
+    enroll,
+    instructor
+};
